basic restructuring untested tho
